# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LUXIMA API Documentation",
        "description": "This is the API documentation for the Luxima project. It covers all endpoints related to authentication and other functionalities.",
        "contact": {
            "email": "admin@luxima.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.luxima.id/api",
            "description": "LUXIMA API Server"
        }
    ],
    "paths": {
...
        "/list-blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get Blog for authenticated user",
                "description": "Retrieve all blog posts for the currently authenticated user.",
                "operationId": "listBlogForAuthUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Blog"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/blog-details/{blog_id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get a single blog for authenticated user",
                "description": "Retrieve the details of a specific blog post for the currently authenticated user.",
                "operationId": "getSingleBlogForAuthUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Blog"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            }
        },
        "/blog/{slug}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get blog by slug",
                "description": "Retrieve the details of a specific blog post by its slug.",
                "operationId": "getBlogBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the blog",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Blog"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            }
        },
        "/Blog/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get a blog by ID",
                "description": "Retrieve the details of a specific blog post.",
                "operationId": "showBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Blog"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            }
        },
        "/Blog": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Create a new blog",
                "description": "Create a new blog post.",
                "operationId": "storeBlog",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlogCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Blog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/blog/{id}": {
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "Update a blog",
                "description": "Update the details of a specific blog post.",
                "operationId": "updateBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlogUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Blog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "summary": "Delete a blog",
                "description": "Delete a specific blog post.",
                "operationId": "destroyBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Blog deleted successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Blog": {
                "title": "Blog",
                "description": "Blog model",
                "required": [
                    "id",
                    "title",
                    "content",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Blog Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Blog content here..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-03T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-03T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "BlogCreateRequest": {
                "title": "Blog Create Request",
                "description": "Request body for creating a blog",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Blog Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Blog content here..."
                    }
                },
                "type": "object"
            },
            "BlogUpdateRequest": {
                "title": "Blog Update Request",
                "description": "Request body for updating a blog",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Updated Blog Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Updated blog content here..."
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}